BIG Changes
Changed logic packages, created filehandler

+++ packages/types/WritePackage.java
@@ -8,7 +8,7 @@ public class WritePackage extends Base_Package implements Package_Executor {
 	private short size;
 	private byte[] file_name;
 
-	public WritePackage(int type, short size, byte[] file_name) {
+	public WritePackage(short type, short size, byte[] file_name) {
 		super(type);
 		this.size = size;
 		this.file_name = file_name;
